requirements.txt
absl-py==2.2.2
astunparse==1.6.3
async-timeout==5.0.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
eventkit==1.0.3
fastdtw==0.3.4
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.55.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
ib-insync==0.9.86
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.9.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.1
motor==3.6.0
mplfinance==0.12.10b0
namex==0.0.8
nest-asyncio==1.6.0
numpy==2.1.3
nvidia-nccl-cu12==2.26.2
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.4
psutil==7.0.0
Pygments==2.19.1
pymongo==4.9.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.8.0
redis==5.2.1
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
ta==0.11.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.0
